fix Devise entity decimal fields precision

+++ src/main/java/fr/microtec/geo2/persistance/entity/tiers/GeoDevise.java
@@ -23,9 +23,9 @@ public class GeoDevise extends ValidateAndModifiedEntity {
 	private String description;
 
 	@Column(name = "dev_tx")
-	private Float taux;
+	private Double taux;
 
 	@Column(name = "dev_tx_achat")
-	private Float tauxAchat;
+	private Double tauxAchat;
 
 }
